+++ src/templates/portfolio-item.jsx
@@ -8,6 +8,8 @@ import Carousel from "../components/Carousel"
 import Button from "../components/Button"
 
 
+
+
 const portfolioItem =  (props) => {
   const {
     description,
@@ -19,7 +21,7 @@ const portfolioItem =  (props) => {
     url,
   } = props.data.item
 
-  
+
 
   return (
     <Layout>
@@ -62,16 +64,17 @@ const portfolioItem =  (props) => {
           </div>
         </div>
       
-      {related && (
+        {related &&  (
         <div className="bg-transparent py-12 lg:py-16 container">
+       
           <div>
             <h2 className="text-3xl sm:text-4xl leading-tight font-extrabold tracking-tight text-stellar mb-8">
               Other projects:
             </h2>
-          </div>
-          <Cards items={related} hideLastItemOnMobile={false} />
+          </div>  
+         <Cards items={related} hideLastItemOnMobile={false} />
         </div>
-      )}
+        )}
